refactoring to put I/O related stuff all in one place
