+++ tests/test_admin.py
@@ -5,7 +5,7 @@ from pathlib import PurePath
 from electionguard.serializable import set_serializers, set_deserializers
 
 from arlo_e2e.admin import make_fresh_election_admin, ElectionAdmin
-from arlo_e2e.utils import write_json_helper, load_json_helper
+from arlo_e2e.ray_io import write_json_helper, load_json_helper
 
 
 class TestAdmin(unittest.TestCase):
