+++ tests/test_ray_io.py
@@ -8,13 +8,14 @@ import coverage
 import ray
 
 from arlo_e2e.ray_helpers import ray_init_localhost
-from arlo_e2e.ray_write_retry import (
+from arlo_e2e.ray_io import (
     set_failure_probability_for_testing,
     write_file_with_retries,
     wait_for_zero_pending_writes,
     reset_pending_state,
+    mkdir_helper,
+    all_files_in_directory,
 )
-from arlo_e2e.utils import mkdir_helper, all_files_in_directory
 
 
 def write_all_files(num_files: int, num_retries: int = 10) -> None:
